Split app mutations
