+++ saleor/graphql/app/mutations/__init__.py
@@ -1,0 +1,27 @@
+from .app_activate import AppActivate
+from .app_create import AppCreate
+from .app_deactivate import AppDeactivate
+from .app_delete import AppDelete
+from .app_delete_failed_installation import AppDeleteFailedInstallation
+from .app_fetch_manifest import AppFetchManifest
+from .app_install import AppInstall
+from .app_retry_install import AppRetryInstall
+from .app_token_create import AppTokenCreate
+from .app_token_delete import AppTokenDelete
+from .app_token_verify import AppTokenVerify
+from .app_update import AppUpdate
+
+__all__ = [
+    "AppActivate",
+    "AppCreate",
+    "AppDeactivate",
+    "AppDeleteFailedInstallation",
+    "AppDelete",
+    "AppFetchManifest",
+    "AppInstall",
+    "AppRetryInstall",
+    "AppTokenCreate",
+    "AppTokenDelete",
+    "AppTokenVerify",
+    "AppUpdate",
+]
