+++ saleor/graphql/app/mutations/app_token_create.py
@@ -1,0 +1,62 @@
+import graphene
+from django.core.exceptions import ValidationError
+from oauthlib.common import generate_token
+
+from ....app import models
+from ....app.error_codes import AppErrorCode
+from ....core.permissions import AppPermission
+from ...account.utils import can_manage_app
+from ...core.mutations import ModelMutation
+from ...core.types import AppError
+from ...utils import get_user_or_app_from_context, requestor_is_superuser
+from ..types import AppToken
+
+
+class AppTokenInput(graphene.InputObjectType):
+    name = graphene.String(description="Name of the token.", required=False)
+    app = graphene.ID(description="ID of app.", required=True)
+
+
+class AppTokenCreate(ModelMutation):
+    auth_token = graphene.types.String(
+        description="The newly created authentication token."
+    )
+
+    class Arguments:
+        input = AppTokenInput(
+            required=True, description="Fields required to create a new auth token."
+        )
+
+    class Meta:
+        description = "Creates a new token."
+        model = models.AppToken
+        object_type = AppToken
+        permissions = (AppPermission.MANAGE_APPS,)
+        error_type_class = AppError
+        error_type_field = "app_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        input_data = data.get("input", {})
+        instance = cls.get_instance(info, **data)
+        cleaned_input = cls.clean_input(info, instance, input_data)
+        instance = cls.construct_instance(instance, cleaned_input)
+        auth_token = generate_token()
+        instance.set_auth_token(auth_token)
+        cls.clean_instance(info, instance)
+        cls.save(info, instance, cleaned_input)
+        cls._save_m2m(info, instance, cleaned_input)
+        response = cls.success_response(instance)
+        response.auth_token = auth_token
+        return response
+
+    @classmethod
+    def clean_input(cls, info, instance, data):
+        cleaned_input = super().clean_input(info, instance, data)
+        app = cleaned_input.get("app")
+        requestor = get_user_or_app_from_context(info.context)
+        if not requestor_is_superuser(requestor) and not can_manage_app(requestor, app):
+            msg = "You can't manage this app."
+            code = AppErrorCode.OUT_OF_SCOPE_APP.value
+            raise ValidationError({"app": ValidationError(msg, code=code)})
+        return cleaned_input
