+++ saleor/graphql/app/tests/mutations/test_app_fetch_manifest.py
@@ -169,7 +169,9 @@ def test_app_fetch_manifest_timeout(
 ):
     mocked_request = Mock()
     mocked_request.side_effect = requests.Timeout()
-    monkeypatch.setattr("saleor.graphql.app.mutations.requests.get", mocked_request)
+    monkeypatch.setattr(
+        "saleor.graphql.app.mutations.app_fetch_manifest.requests.get", mocked_request
+    )
     manifest_url = "http://localhost:3000/manifest-doesnt-exist"
     query = APP_FETCH_MANIFEST_MUTATION
     variables = {
